Decode arguments
This will be completly overworked, because it's more simple to use a buffered reader instead of passing arrays/slices and manually change array positions.
Converting the raw message bytes to the specified type can be done via transmuting, but this is an unsafe operation. Another disadvantage of this approach is that I have to implement endianess conversion by myself.