Test OSC True, False, Nil and Infinity
Maybe, test is not the right word. There is no direct assignement
between type tags and decoded arguments, so I can't check if the boolean
value should be true or false, at least not as long as I use one bool
type and not True and False seperately like in the OSC 1.0 specification.
The lib chan choose what types Nil and Inf should have, so there is also
no real test for these OSC arguments.