Make `encode_message` public
Creating an `OscPacket::Message(msg)` from a predefined message will
move the ownership because the `Copy` trait is not implemented for the
`OscMessage` type. This means that you get `use of moved value` errors
if you try to access values of `msg` after you stuffed it in an `OscPacket`.
Most times it's ok to move the value, but especially for testing it
would be nice to check against the original fields.
Using references for the `OscPacket::Message` argument is not a
clean solution because it will introduce explicit lifetimes for the
`OscPacket` enum as well as the `OscBundle` struct. This alone would be
tolerable, but lifetimes have to be assigned for `decoder::decode` too, because
it returns an `Result<OscPacket>`.