OscTime -> SystemTime conversion: round nanos.
The fractional part of OscTime is converted to a f64 number
of nanoseconds, but Duration::new expects an u32 for its nanos
component.

By rounding to the nearest integer number of nanoseconds
(as is already done in OscTime::try_from) instead of truncating down,
we make the conversion slightly more accurate, especially on Windows
where SystemTime only has a 100ns resolution.

(in the longer term, we'll probably want to round to the nearest 100ns
on Windows)