OscTime conversions tests: greater tolerance on Windows
On Windows, the resolution of SystemTime is 100ns, as opposed to 1ns on UNIX
(https://doc.rust-lang.org/std/time/struct.SystemTime.html#platform-specific-behavior).

As a result, any conversion of OscTime to SystemTime results in the
latter being quantized to the nearest 100ns (rounded down).
This also means both types of round-trips are lossy: for any time `t`,
`t + i` roundtrips to `t` for all `0ns <= i < 100ns`.

This commit increases the tolerance for equality *on Windows only*
to 100ns instead of 5 to compensate for that, which in turn
makes the unit tests pass on Windows.