Rework encoder tests
This makes the naming consistent by using packet as the name for the
OSC packet to be encoded, and bytes as the name for vector of bytes that
contains the encoded packet.  Besides renaming of variables this also
simplifies the identity check by not unwrapping the packet to compare
the contents, and instead asserting that the packet itself is equal.