Fix decoding of four-byte aligned strings
This fixes a bug that occurred when the content-length of a string,
excluding the null-byte, was four-byte aligned. In this case we added 4
additional bytes of padding, instead of 3. We now consume the string
including the zero-byte in the parser, and then apply the correct
null-byte padding.

Thanks to @edjin for providing the initial fix this patch is based on.